rust, rustup: Use MSVC by default
Changes rust and rustup packages to use MSVC toolchain by default, as
this is generally the better supported (and default) choice.
In the rustup package, the MSVC version is always used (the GCC build
of rustup provides no benefits, and the MSVC version can be used to
install MSVC toolchains). The --default-host flag has been removed, so
new installs will use the rustup default (MSVC) but existing installs
will have a configuration file indicating GCC is the default; a note was
added to the rustup packages stating how to change this.
The rust-gnu package was added, to allow using the GCC toolchain without
rustup, and a note was added to all rust packages indicating that rustup
can be used to more easily manage multiple toolchains (and is the
primary install method; see
https://forge.rust-lang.org/infra/other-installation-methods.html).

Closes #931.